Clippy pedantic

+++ src/decoders/json_types/heading.rs
@@ -54,8 +54,8 @@ impl fmt::Display for Heading {
         match self {
             // cast to u32 to remove the decimal
             Heading::HeadingAsInteger(heading) => write!(f, "{heading} degrees"),
-            Heading::HeadingAsFloat(heading) => write!(f, "{} degrees", *heading as u32),
-            Heading::HeadingAsFloat64(heading) => write!(f, "{} degrees", *heading as u32),
+            Heading::HeadingAsFloat(heading) => write!(f, "{} degrees", *heading),
+            Heading::HeadingAsFloat64(heading) => write!(f, "{} degrees", *heading),
             Heading::None => write!(f, "None"),
         }
     }
